obexd: Add initial implementation of mimetype driver and filesystem plugin.
Mimetype driver handles object access if there is no driver for the given
mimetype the default one (NULL) is used, which is file driver for instance.

+++ obexd/src/opp.c
@@ -107,7 +107,7 @@ void opp_get(obex_t *obex, obex_object_t *obj)
 {
 	struct obex_session *os;
 	obex_headerdata_t hv;
-	guint32 size;
+	size_t size;
 
 	os = OBEX_GetUserData(obex);
 	if (os == NULL)
